requirements.txt
torch>=2.0
torchvision
pytorch-lightning>=2.0
opencv-python
tqdm
pillow
numpy
albumentations==1.4.17
segment-anything